Revamp action and undo/redo
The undo crashed pointed to a fundamental bug in the action and
undo/redo architecture: The actions were created relative to a
workspace, instead of the SuperNode to which they belong, and the
model and view changes were mixed. Now, the WorkSpace only modifies
view and the rest is directly done in the actions, and if the WorkSpace
shows a different SuperNode than the action, the view is not changed.
Thus, we can have undo/redo outside of the current view, which still
must be communicated to the user.